+++ src/main/java/ru/yandex/practicum/filmorate/repository/FilmRepository.java
@@ -45,7 +45,7 @@ public interface FilmRepository {
      * @param grade  Оценка фильма
      * @return Фильм с обновленным списком лайков
      */
-    Film putLike(Integer id, Integer userId, Integer grade);
+    Film putLike(Integer id, Integer userId, Double grade);
 
     /**
      * Удаление лайка пользователя
